+++ adapters/PubMatic/PubMaticAdapter/NativeAdLoader.swift
@@ -48,6 +48,14 @@ final class NativeAdLoader: NSObject {
   }
 
   func loadAd() {
+    if adConfiguration.bidResponse != nil {
+      loadRTBAd()
+    } else {
+      loadWaterfallAd()
+    }
+  }
+
+  private func loadRTBAd() {
     guard let bidResponse = adConfiguration.bidResponse, let watermark = adConfiguration.watermark
     else {
       handleLoadedAd(
@@ -59,6 +67,18 @@ final class NativeAdLoader: NSObject {
       return
     }
     client.loadRtbNativeAd(bidResponse: bidResponse, delegate: self, watermarkData: watermark)
+  }
+
+  private func loadWaterfallAd() {
+    do {
+      let publisherId = try Util.publisherId(from: adConfiguration)
+      let profileId = try Util.profileId(from: adConfiguration)
+      let adUnitId = try Util.adUnitId(from: adConfiguration)
+      client.loadWaterfallNativeAd(
+        publisherId: publisherId, profileId: profileId, adUnitId: adUnitId, delegate: self)
+    } catch {
+      handleLoadedAd(nil, error: error.toNSError())
+    }
   }
 
   private func handleLoadedAd(_ ad: MediationNativeAd?, error: Error?) {